NXDRIVE-695: File is not deleted on the server when the parent folder is renamed, while offline
* Clean-up
* Add test
* Update all DB local paths on FS move event

This fixes the issue on macOS and Windows.

On GNU/Linux, the deleted event is against the old
"/folder/subfolder/file.txt" instead of
"/folder-renamed/subfolder-renamed/file.txt".

This is a corner case and I have no idea how to fix it.
So the fix is postponed on GNU/Linux.
To follow with NXDRIVE-1690.

* Use int() in BlackListQueue

To restore old behavior and fix randoms failures in
test_blacklist_queue.py.

+++ nxdrive/engine/blacklist_queue.py
@@ -14,10 +14,10 @@ class BlacklistItem:
         self._item = item
         self._interval = next_try
 
-        self._next_try = self._interval + time.monotonic()
+        self._next_try = self._interval + int(time.monotonic())
         self.count = 1
 
-    def check(self, cur_time: float) -> bool:
+    def check(self, cur_time: int) -> bool:
         return cur_time > self._next_try
 
     def get(self):
@@ -38,7 +38,7 @@ class BlacklistQueue:
             self._queue[item.uid] = item
 
     def get(self) -> Generator[BlacklistItem, None, None]:
-        cur_time = time.monotonic()
+        cur_time = int(time.monotonic())
         with self._lock:
             for item in self._queue.copy().values():
                 if item.check(cur_time):
